% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{clean_spikes}
\alias{clean_spikes}
\title{Remove strong additive outliers, i.e. "spikes"}
\usage{
clean_spikes(x)
}
\arguments{
\item{x}{A vector of time-ordered observations}
}
\value{
An adjusted vector of observations in which spieks have been replaced with running median
}
\description{
Remove strong additive outliers, i.e. "spikes"
}
