% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis.R
\name{cluster_func}
\alias{cluster_func}
\title{Cluster time series, plot dendrogram, print cluster validation indices}
\usage{
cluster_func(x, type = "hierarchical", distance = "gak", preproc = NULL,
  trace = TRUE, control = "ward.D2")
}
\arguments{
\item{x}{A data frame of row-wise time series}

\item{type}{See dtwclust::tso}

\item{distance}{See dtwclust::tso}

\item{preproc}{See dtwclust::tso}

\item{trace}{See dtwclust::tso}

\item{control}{See dtwclust::tso}
}
\value{
A cluster object, dendrogram, and cluster validation indices
}
\description{
Cluster time series, plot dendrogram, print cluster validation indices
}
