% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis.R
\name{clusterprep}
\alias{clusterprep}
\title{Convert data to aligned XTS object and transpose}
\usage{
clusterprep(x, omit)
}
\arguments{
\item{x}{A tibble or data frame}

\item{omit}{A vector of strings indicating ScaleIDs to be dropped}
}
\value{
A data frame of row-wise time series
}
\description{
Convert data to aligned XTS object and transpose
}
