% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{data_proc}
\alias{data_proc}
\title{\code{data_proc} generates and appends four new fields: a 25 h running mean of the despiekd weights,
  a loess smoothing of the 25-h running mean,
  a first-order difference vector of the loess-smoothed 25-h running mean,
  and a z-normalization of the difference vector.}
\usage{
data_proc(x)
}
\arguments{
\item{x}{A tibble or data frame}
}
\value{
A tibble with added fields
}
\description{
\code{data_proc} generates and appends four new fields: a 25 h running mean of the despiekd weights,
  a loess smoothing of the 25-h running mean,
  a first-order difference vector of the loess-smoothed 25-h running mean,
  and a z-normalization of the difference vector.
}
