% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{deartifact}
\alias{deartifact}
\title{\code{deartifacr} corrects artifacts by subtacting from each observation the cumulative sum of observations > delta_max.
Empirically, histograms of my differenced show a consistent discontinuity around 2.5, suggesting that this is an appropriate value for delta_max.}
\usage{
deartifact(x, delta_max = 2.5)
}
\arguments{
\item{x}{A vector of differenced time-ordered observations}

\item{delta_max}{A real number defining the weight chenge threshold above which a weight change is considered an artifact.}
}
\value{
A vector of de-artifacted observations
}
\description{
\code{deartifacr} corrects artifacts by subtacting from each observation the cumulative sum of observations > delta_max.
Empirically, histograms of my differenced show a consistent discontinuity around 2.5, suggesting that this is an appropriate value for delta_max.
}
