% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{fix_outliers}
\alias{fix_outliers}
\title{Use ARIMA modeling to identify and correct additive and level-shift outliers; one hive at a time}
\usage{
fix_outliers(x, ID, cval = 9, method = "arima")
}
\arguments{
\item{x}{A data frame or tibble}

\item{ID}{A string corresponding to the ScaleID field, specifying which hive is to be processed}

\item{cval}{A real number, parameter of the tso function modulating sensitivity}

\item{method}{A string specifying the modeling method to be used to identify outliers}
}
\value{
A tibble containing outlier-corrected vector of observations
}
\description{
Use ARIMA modeling to identify and correct additive and level-shift outliers; one hive at a time
}
