% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{fix_outliers_batch}
\alias{fix_outliers_batch}
\title{Use ARIMA modeling to identify and correct additive and level-shift outliers; batch process multiple hives}
\usage{
fix_outliers_batch(x, cval = 9, method = "arima")
}
\arguments{
\item{x}{A data frame or tibble}

\item{cval}{A real number, parameter of the tso function modulating sensitivity}

\item{method}{A string specifying the modeling method to be used to identify outliers}
}
\value{
A tibble containing outlier-corrected vector of observations
}
\description{
Use ARIMA modeling to identify and correct additive and level-shift outliers; batch process multiple hives
}
