% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{get_names}
\alias{get_names}
\title{Split HiveName field into Site and Hive#; assumes HiveName follows format site.#}
\usage{
get_names(FilePath)
}
\arguments{
\item{FilePath}{Path to BroodMinder SQLite database}
}
\value{
A tibble consisting of fields UUID, ScaleID, Site, Hive, filtered to contain only unique rows
}
\description{
Split HiveName field into Site and Hive#; assumes HiveName follows format site.#
}
