% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis.R
\name{inout}
\alias{inout}
\title{\code{inout} creates a new field in the tabulated distance matrix specifiying whether each pairwise comparison is of hives within a site or hives ar different sites}
\usage{
inout(x, y)
}
\arguments{
\item{x, y}{Columns in tabulated distance matrix representing each of the two hives in each pairwise comparison}
}
\description{
\code{inout} creates a new field in the tabulated distance matrix specifiying whether each pairwise comparison is of hives within a site or hives ar different sites
}
