% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{load_bm}
\alias{load_bm}
\title{\code{load_bm} loads a BroodMinder SQLite database and extracts logged readings into a tibble.}
\usage{
load_bm(FilePath, tz, ftrim, btrim, sep = "\\\\.")
}
\arguments{
\item{FilePath}{Path to BroodMinder SQLite database}

\item{tz}{String identifying time zone}

\item{ftrim}{String "YYYY-mm-dd" indicating front date cutoff}

\item{btrim}{String "YYYY-mm-dd" indicating back date cutoff}

\item{sep}{A string denoting the separator between the site and hive number in the HiveName field; assumes HiveName follows format site % sep % hive#}
}
\value{
A tibble containing selected fields and trimmed to front and back cutoff dates
}
\description{
\code{load_bm} loads a BroodMinder SQLite database and extracts logged readings into a tibble.
}
