% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{load_db}
\alias{load_db}
\title{Load BroodMinder SQLite database and extract logged readings into a tibble}
\usage{
load_db(FilePath, tz, ftrim, btrim)
}
\arguments{
\item{FilePath}{Path to BroodMinder SQLite database}

\item{tz}{String identifying time zone}

\item{ftrim}{String "YYYY-mm-dd" indicating front date cutoff}

\item{btrim}{String "YYYY-mm-dd" indicating back date cutoff}
}
\value{
A tibble containing selected fields and trimmed to front and back cutoff dates
}
\description{
Load BroodMinder SQLite database and extract logged readings into a tibble
}
