% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{mean25h}
\alias{mean25h}
\title{\code{mean25h} removes daily periodicity by honey bee weight data by taking a 25-h rolling mean}
\usage{
mean25h(x)
}
\arguments{
\item{x}{A vector of time-ordered observations}
}
\value{
A vector of 25-h means with NA pad for first 12 and last 12 observations
}
\description{
\code{mean25h} removes daily periodicity by honey bee weight data by taking a 25-h rolling mean
}
