% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{plot_delta}
\alias{plot_delta}
\title{\code{plot_delta} is a wrapper for ggplot that allows the plotting of differenced weight metrics by hive, by site, or pooled}
\usage{
plot_delta(x, type = "points", by = "none", metric, knots = 25,
  omit = NULL)
}
\arguments{
\item{x}{a tibble output by data_proc}

\item{type}{a string specifying the type of plot; options are "points" for scatterplot or "hist" for histogram; note that for histograms, xlim is set to c(-5, 5), which results in the trimming of extreme values but improves visualization of data. Similarly, for scatterplots, ylim is set to (-0.15, 0.15), which trims extreme values but improves visualization of the smooth fit line}

\item{by}{a string specifying whether plot should be faceted by "site", "scale", or "none"}

\item{metric}{a string vector indicating which weight metrics should be plotted}

\item{knots}{a real number controlling the number of knots in the GAM splines smoothing function; lower number = more smoothing}

\item{omit}{a string indicating which ScaleIDs shoud be dropped prior to plotting}
}
\value{
a plot
}
\description{
\code{plot_delta} is a wrapper for ggplot that allows the plotting of differenced weight metrics by hive, by site, or pooled
}
