% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{plot_wt}
\alias{plot_wt}
\title{\code{plot_wt} is a wrapper for ggplot that allows the plotting of weight metrics by hive, by site, or pooled}
\usage{
plot_wt(x, by = "site", metric, knots = 25, omit = NULL)
}
\arguments{
\item{x}{a tibble output by data_proc}

\item{by}{a string specifying whether plot should be faceted by "site", "scale", or "none"}

\item{metric}{a string vector indicating which weight metrics should be plotted}

\item{knots}{a real number controlling the number of knots in the GAM splines smoothing function; lower number = more smoothing}

\item{omit}{a string indicating which ScaleIDs shoud be dropped prior to plotting}
}
\value{
a plot
}
\description{
\code{plot_wt} is a wrapper for ggplot that allows the plotting of weight metrics by hive, by site, or pooled
}
