% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis.R
\name{site_effect}
\alias{site_effect}
\title{\code{site_effect} draws boxplots and runs t.test to compare pairwise distance withing and across sites}
\usage{
site_effect(x)
}
\arguments{
\item{x}{a tabulated distance matrix}
}
\description{
\code{site_effect} draws boxplots and runs t.test to compare pairwise distance withing and across sites
}
