% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis.R
\name{tab_dist}
\alias{tab_dist}
\title{Convert distance matrix to long-form tibble in prep for downstream analysis}
\usage{
tab_dist(x)
}
\arguments{
\item{x}{A dtwclust object containing the slot distmat}
}
\value{
A longform tibble with pairwise distances by site
}
\description{
Convert distance matrix to long-form tibble in prep for downstream analysis
}
