% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis.R
\name{xts_align}
\alias{xts_align}
\title{Align time series to nearest hour}
\usage{
xts_align(x)
}
\arguments{
\item{x}{An XTS time series object}
}
\value{
An XTS time series object with observations aligned to nearest hour
}
\description{
Align time series to nearest hour
}
