% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{clean_spikes}
\alias{clean_spikes}
\title{\code{clean_spikes} removes strong additive outliers, i.e. "spikes", using the \code{despike} function from package \code{oce}. Parameters are set so that strong spike values are replaced with a running median of width = 7}
\usage{
clean_spikes(x)
}
\arguments{
\item{x}{A vector of time-ordered observations}
}
\value{
An adjusted vector of observations in which spieks have been replaced with running median
}
\description{
\code{clean_spikes} removes strong additive outliers, i.e. "spikes", using the \code{despike} function from package \code{oce}. Parameters are set so that strong spike values are replaced with a running median of width = 7
}
