% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{data_proc}
\alias{data_proc}
\title{Apply 25 h running mean, loess smoothing, differencing, and scaling in preparation for downstream analysis}
\usage{
data_proc(x)
}
\arguments{
\item{x}{A tibble or data frame}
}
\value{
A tibble with added fields
}
\description{
Apply 25 h running mean, loess smoothing, differencing, and scaling in preparation for downstream analysis
}
